+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 (IN PROGRESS)
 
 * Specify module name in settings second column. Part of STRPCOMP-1.
+* Include label in loan-policies editor paneTitle. Part of STRPCOMP-1.
 
 ## [1.0.0](https://github.com/folio-org/ui-scan/tree/v1.0.0) (2017-07-03)
 [Full Changelog](https://github.com/folio-org/ui-scan/compare/v0.5.0...v1.0.0)