+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Specify module name in settings second column. Part of STRPCOMP-1.
 * Include label in loan-policies editor paneTitle. Part of STRPCOMP-1.
 * Eliminate use of `componentWillMount` to provide initial values to resources. Part of STRIPES-433.
-* Improvements to loan period form in settings. Fixes UIS-36, UIS-38. 
+* Improvements to loan period form in settings. Fixes UIS-36, UIS-38.
+* Rewrite nav menu selector in loan policy form to be reusable. Also fixes UIS-39.
 
 ## [1.0.0](https://github.com/folio-org/ui-scan/tree/v1.0.0) (2017-07-03)
 [Full Changelog](https://github.com/folio-org/ui-scan/compare/v0.5.0...v1.0.0)